Add files via upload
This contains the algorithm I made for the elevator to decide which direction to go. The third while loop is the latest version of the code, comment out the other while loops if you just want to run the latest version. I plan to improve the annotations later as they're fairly barebones currently